+++ resto-code-front/src/Components/RUTAS/Rutas.jsx
@@ -1,4 +1,4 @@
-import Home from "../../Pages/Home";
+import Home from "../../Pages/HOME/Home";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 
 const Rutas = () => {
