+++ resto-code-front/src/Pages/HOME/Home.css
@@ -1,0 +1,304 @@
+@import url("https://fonts.googleapis.com/css2?family=Bebas+Neue&display=swap");
+.contendor-botonReserva-imagenes {
+  height: 25rem;
+  display: flex;
+}
+.contenedorBotonReservar {
+  width: 50%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  background-color: #181818;
+}
+.tituloRestoCode {
+  font-family: "Bebas Neue", sans-serif;
+  color: white;
+  display: flex;
+  flex-direction: column;
+  text-align: center;
+}
+.tituloRestoCode h1 {
+  font-size: 4rem;
+  padding: 0;
+  margin: 0;
+}
+.restoBar h2 {
+  border-top: 2px solid #68b2f8;
+  border-bottom: 2px solid #68b2f8;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding-top: 4px;
+}
+.botonReservarIndex {
+  border: none;
+  background-color: #1d0c20;
+  color: white;
+  padding: 6px 10px 6px 10px;
+  border: 2px solid #651f71;
+  transition: 0.2s;
+  border-radius: 3px;
+  font-size: 1.5rem;
+  margin-top: 1rem;
+  font-family: "Bebas Neue", sans-serif;
+}
+.botonReservarIndex:hover {
+  background-color: #7037cd;
+  border: none;
+  margin: 0;
+}
+.imagenDeFondoIndex {
+  width: 50%;
+  background-image: url(../../Image/INDEX-IMAGE/fondoIndex7.jpg);
+  background-size: cover;
+  background-repeat: no-repeat;
+}
+.sombraImagenInicioIndex {
+  background-color: rgba(0, 0, 0, 0.411);
+  width: 100%;
+  height: 100%;
+}
+.imagenDeFondoEnResponsive {
+  width: 100%;
+  background-image: url(../../Image/INDEX-IMAGE/fondoIndex7.jpg);
+  background-size: cover;
+  background-repeat: no-repeat;
+  display: none;
+}
+.contendorFrases {
+  /* border: 1px solid rgb(211, 214, 16); */
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  height: 9.5rem;
+}
+.tituloDePresentacionMenu {
+  font-family: "Bebas Neue", sans-serif;
+  font-size: 3rem;
+  color: #181818;
+}
+.mensajeConciencia {
+  font-family: "Bebas Neue", sans-serif;
+  font-size: 1.8rem;
+  color: #7037cd;
+}
+.contendorCerveza {
+  background-color: #181818;
+  display: flex;
+}
+.cartaMenuDeCervezas {
+  width: 50%;
+}
+.cajaConFrase {
+  width: 50%;
+  display: flex;
+  flex-direction: column;
+  align-items: center !important;
+  justify-content: center;
+  font-family: "Bebas Neue", sans-serif;
+}
+.cajaConFrase h2 {
+  font-size: 3.5rem;
+  color: white;
+}
+.cajaConFrase h3 {
+  font-size: 2.5rem;
+  color: #7037cd;
+}
+.imagenDeAmbosMenus {
+  height: 40rem;
+  display: flex;
+  background-image: url(../../Image/INDEX-IMAGE/fondoDeMenus2.jpeg);
+  background-size: cover;
+  background-repeat: no-repeat;
+  /* border: 1px solid seagreen; */
+}
+.menuUno {
+  width: 50%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  padding: 0.5rem;
+  background-color: rgba(0, 0, 0, 0.411);
+  /* border: 1px solid rgb(105, 46, 139); */
+}
+.menuUno img {
+  height: 100%;
+}
+.menuUno img:hover {
+  padding: 1.5rem;
+}
+.menuDos {
+  width: 50%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  padding: 0.5rem;
+  background-color: rgba(0, 0, 0, 0.411);
+  /* border: 1px solid rgb(105, 46, 139); */
+}
+.menuDos img {
+  height: 100%;
+}
+.menuDos img:hover {
+  padding: 1.5rem;
+}
+.contendorContacto {
+  /* border: 1px solid rgb(105, 46, 139); */
+  display: flex;
+  flex-direction: column;
+  text-align: center;
+  font-family: "Bebas Neue", sans-serif;
+  padding-top: 2rem;
+  padding-bottom: 1.5rem;
+}
+.contendorContacto h2 {
+  color: #181818;
+}
+.contendorContacto h3 {
+  color: #7037cd;
+}
+.contendorDeMapaRedesSociales {
+  height: 25rem;
+}
+.cajaMapa {
+  height: 100%;
+  display: flex;
+  /* border: 1px solid saddlebrown; */
+  border-radius: 6px !important;
+}
+.mapaContacto {
+  width: 50%;
+  height: 100%;
+  padding: 2rem;
+  /* border: 1px solid saddlebrown; */
+}
+.contenedorRedesContacto {
+  width: 50%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  /* border: 1px solid seagreen; */
+}
+.cajaQrWp {
+  width: 10rem;
+  height: 10rem;
+  margin-bottom: 10px;
+  /* border: #68b2f8 1px solid; */
+}
+.cajaQrWp img {
+  width: 100%;
+  height: 100%;
+}
+.cajaQrIg {
+  width: 12.5rem;
+  height: 12.5rem;
+  margin-bottom: 10px;
+  /* border: #68b2f8 1px solid; */
+  background-image: url(../../Image/INDEX-IMAGE/restocodeokqr.png);
+  background-size: cover;
+  background-repeat: no-repeat;
+}
+
+/*------------------ RESPONSIVE ------------------*/
+/* Iphone 12 Pro Max */
+@media (min-width: 360px) and (max-width: 914px) {
+  .contenedorBotonReservar {
+    display: none;
+  }
+  .imagenDeFondoIndex {
+    display: none;
+  }
+  .imagenDeFondoEnResponsive {
+    display: block;
+  }
+  .sombraImagenInicioIndex {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+  }
+  .tituloRestoCoderResponsive {
+    font-family: "Bebas Neue", sans-serif;
+    color: white;
+  }
+  .tituloRestoCoderResponsive h1 {
+    font-size: 3.5rem;
+  }
+  .restoBarResponsive h2 {
+    border-top: 2px solid #68b2f8;
+    border-bottom: 2px solid #68b2f8;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding-top: 4px;
+    font-size: 2.5rem;
+  }
+  .botonReservarIndexResponsive {
+    border: none;
+    background-color: #1d0c20;
+    color: white;
+    padding: 6px 10px 6px 10px;
+    border: 2px solid #651f71;
+    transition: 0.2s;
+    border-radius: 3px;
+    font-size: 1.5rem;
+    margin-top: 1rem;
+    font-family: "Bebas Neue", sans-serif;
+  }
+  .contendorFrases {
+    height: 8rem;
+  }
+  .tituloDePresentacionMenu {
+    font-size: 2rem;
+  }
+  .imagenDeAmbosMenus {
+    height: 50rem;
+    flex-direction: column;
+    align-items: center;
+  }
+  .menuUno {
+    width: 100%;
+    height: 50%;
+  }
+  .menuDos {
+    width: 100%;
+    height: 50%;
+  }
+  .contendorCerveza {
+    flex-direction: column;
+  }
+  .cartaMenuDeCervezas {
+    width: 100%;
+    height: 50%;
+  }
+  .cajaConFrase {
+    width: 100%;
+    height: 50%;
+  }
+  .cajaConFrase h2 {
+    font-size: 2.5rem;
+  }
+  .cajaConFrase h3 {
+    font-size: 2rem;
+  }
+  .cajaMapa{
+    flex-direction: column;
+  }
+  .mapaContacto {
+    width: 100%;
+    height: 50%;
+    padding: 2rem;
+  }
+  .contenedorRedesContacto {
+    flex-direction: row;
+    height: 50%;
+    width: 100%;
+    padding-left: 1.5rem;
+  }
+ 
+}
