--- resto-code-front/src/Pages/Home.jsx
@@ -1,10 +1,0 @@
-import ContenidoIndex from "../Components/CONTENIDO INDEX/ContenidoIndex"
-const Home = () => {
-  return (
-    <>
-    <ContenidoIndex />
-    </>
-  )
-}
-
-export default Home
